requirements.txt
# =========================
# Deep Learning (CUDA)
# =========================
torch==2.7.1+cu118
torchvision==0.22.1+cu118
torchaudio==2.7.1+cu118

# =========================
# Data & Images
# =========================
numpy==1.26.4
pandas==2.3.3
Pillow==10.4.0
openpyxl==3.1.5

# =========================
# Training & Evaluation
# =========================
scikit-learn==1.8.0
tqdm==4.67.1
matplotlib==3.10.8
seaborn==0.13.2

# =========================
# Dashboard
# =========================
streamlit==1.32.0
